requirements.txt
sounddevice
numpy
adafruit-circuitpython-neopixel